Updated Rdf Transform to get and parse build phases from Shipweight.
Background: In Shipweight equipment has a phase code. In the Code table phase codes (generic identifier C02) give a "normalized" start and stop time.
The Item_Filter table provides, among other things, a "Built-As" filters with a cut-off time. In other words, if start <= time and stop >= time
then the item was in place when the platform was built.

Expanded the transformation logic for Shipweight to include
 - reading the Code and Item_Filter tables
 - transforming the result to triples (the relevant filter is type as facility:AsBuilt)

Updated the transformation to take a series of input DataTables as a DataSet

Fixed a bug in the Shipweight provenance information that swapped the prov:wasGeneratedBy and prov:wasDerivedFrom predicates.