fix: handle underscore (#104)
* chore: refactored tests and added case for `_`

* chore: remove unused variables

* fix: Separate base Uri in own parameter

Solved `_`-problem by using base as a separate parameter. Before this
was done by splitting on `_`. For example:

when we where to attach property on
https://rdf.equinor.com/test/linelist/20L00015A_WallThicknessDatum we
would access the previous property
https://rdf.equinor.com/test/linelist/20L00015A_WallThickness, split on
`_` to get base and attach the new type WallThicknessDatum. This would
of course create problem if `_` appeared anywhere in the Uri.

New solution is to pass base uri as parameter all the way through the
recursion in addition to the subjectIndividual we are currently working
on.