chore: add endpoint for information (#132)
* chore: add endpoint for information

Add info endpoint in splinter. Should be used for debugging and
dashboard

Hit some obstacles when configuring for local development. Add readme +
exception and removed appsettings.Development.json

* Update bicep and server setup

Update bicep and server setup with new object model for configuration.

* Try to handle array in appsettings in bicep

When configuring appsettings in bicep we use name and value pairs. If we
want to use nested settings like
```
"a": {
    "b": "c"
}
```
we must specify name like `"a__b"` (nesting becomes `__`). However I
could not find any good way of specifying the value as an array. If we
specify the array directly in value we get:

> HTTP request body must not be empty.

Closest thing I found to a solution was:

https://stackoverflow.com/questions/65790593/azure-app-services-add-array-of-strings-in-configuration

In this commit I am testing wrapping array in a string()-function

* The same way __ can be used to traverse the json tree in appsettings if we write a__0 and a__1 it means that the value of a is an array and 0 and 1 is the first and second element respectively. This is a bit strange, but it works. See https://stackoverflow.com/questions/34063167/using-an-array-in-azure-web-app-settings