feat: new empty endpoints for revisioned data (#136)
* feat: new endpoints

New Endpoints in DataController and RevisionTrainController. Not yet implemented, but the plan is to start working on them both from an admin tool client and actual implementation at the same time.

Hopefully the new structure with revision trains that groups revision of data will make it easier to visualize what kind of data that is available at all the fusekis that splinter is controlling. Splinter can now synchronize records about data when uploading new data. For example if we upload a specific revision of mel excel, splinter's main fuseki will also add a new reference to this revision. This reference includes the name of the fuseki holding these data, and the named graph. Listing these pairs of servers and named graphs can let the client easely delete unwanted data sets.

* Changes after talk with Bård.

Instead of flexible revisions trains we now hard code Ens revision trains to a fixed list of parameters. These are:

* DocumentName (also used as ID)
* ProjectCode
* FacilityName
* ContractNumber
* DocumentType (Mel / Line List or SCD etc.)