fix: Remove hardcoded server from Provenance (#140)
* fix: Remove hardcoded server from Provenance

"Dugtrio" was used as a hardcoded Fuseki name as a temporary solution
before Splinter's routing is available. This caused issues for new functionality,
like retrieving data from CommonLib.

The major changes in this PR are groundwork for continued improvement of Splinter's
codebase as the new APIs are implemented:
- The dependencies on hardcoded server names has either been removed or
moved to the highest entrypoint into Splinter, i.e. TieMelConsumer still needs a server name.
- The methods for getting previous revisions have been moved up the stack
and is now called from without the ProvenanceService itself. The idea is to
clarify the role of the FusekiService and going forward reduce the number of relations
to other services.

Minor changes:
- Removed option to get previous revisions for data not stored in named graphs
- Minor changes to logging and log levels of some messages.

* chore: Update main Fuskei name from meta to main