feat: Create named graph (#163)
* chore: Refactoring Fuseki services

* chore: Add exception handling for minimal API endpoints

* chore: Refactor Fuseki service

- Updated minimal API for revision train endpoints to return Fuseki result
- Renamed Fuseki responses
- Merged Fuseki's Ask and Select responses into FusekiQueryService
- Updated queries in RevisionTrainService to return rdf
- Removed obsolete MVC Exception filter

* feat: Migrate form-data endpoint to Minimal API

Scaffolding for a single POST endpoint has been added for Named Graphs.
Named graphs are posted as files, and the endpoint accepts AML, Excel
and Turtle.

Removed some of the obsolete endpoints.

* chore: Removed application/octet-stream from supported content types

* feat: Rework provenance - WIP

With the exception of revision name and revision date, provenance info
is kept at the revision train. The revision train is retrieved and there
is no longer a need for provenance information to go through a Datatable.

Updated shacl to require a splinter:trainType. This concept will evolve
as new revision train types emerges.

* chore: Re-wireing the mel-excel-transformation flow. WIP

* chore: Prepare validation of incoming revision

Whether an incoming dataset is a valid revision will be checked before
the actual transformation of data. WIP

* chore: Parse train graph to RevisionTrainModel

The revision train graph is parsed into the RevisionTrainModel.
WIP - Add parsing of the list of named graphs.

* chore: Update graph parser to include named graphs

* feat: Add initial record context generation

Note: "Named graphs" will be referred to as "records" going forward.
This is a change that haven't reached the code yet.

Prior to upload a record the following ground-work is done:
- The revision train graph is parsed into a revision train class model
- The new revision is validated, i.e. checks are performed to ensure that
    the uploaded data is a new revision.
- A a new revision context of the record is created.

For the actual transformation step (NamedGraphService.UploadExcel):
- the pipeline has been given duplicate methods to transform the raw data

Still to be done in PR:
- post the transformed data to the correct triple store as a named graph
- post the revision context to the revision train in the "main" triple store.

* feat: Post RecordContext and Record to Fuseki

Both the RecordContext and Record are now posted to the correct Fusekis.
A couple of additional triples in RecordContext (still called NamedGraph
in code) is needed before all the pieces are in place.

Reenabled code to run tests. Tests needs an update.

* chore: Change update endpoint Date type to string

* feat: Add records to train and as named graphs

Records are added as entries to the revision trains and their content is
added in named graphs.

* chore: Update tests and remove old code

- Tests are updated to run against revision train
- Started removing old code. This is ongoing work as more and more of the
old functionality is ported onto the new endpoints.
- Started renaming Named Graphs to Records.
- Fix in ASK query to check whether a train already exist.

* chore: Fixed final conflict

* chore: Removed LocalFunction project from sln

* fix sku

* chore: Update from code review feedback.

- Fix issue with not creating record:replace for revisions > 2
- Continued renaming named graphs to records.
- Fix async without await

Co-authored-by: baard-henning <batv@equinor.com>
Co-authored-by: toftalisk <martin.tofteberg@gmail.com>