feat: Transform data using ontologies (#175)
Previous transformations have been performed raw, i.e. the data are
transformed to triples using the source predicates. Now, the source triples
can be further transformed into the nomenclature of an ontology. At the moment,
the transformation returns data represented in different levels of complexity
based on the ontology used. This can in the future be made more configurable,
for instance the user can decide whether the transformation should be basic or complex.

  Changes:
    - Updated various classes/methods to return DotnetRdf.Graph.
    - Updated the conversion logic work on a graph instead of datatable
    - Various changes to generalize the conversion
    - Updated test cases
    - A bit of cleaning up and a few minor bug fixes