ci: add csproj sync (#21)
A new github action that syncs content from release please version.txt into the csproj file that is used to create nuget packages. The action should trigger on pushes to the release please branch and also commit on that branch. This way both release notes and the version will be updated in the same release please PR