fix: Update excel reader to correctly parse rows
ExcelDomReaderService use input parameters to select the relevant data rows.
In addition to input, Rows.Count() was used to get an upper limit on the number of rows.
The counted number of rows is not the same as the maximum row index,
but rather the number of rows with actual content. This made the selection algorithm
fail for spreadsheets with empty rows.

The solution is to filter data based on the RowIndex, which match the actual
row displayed in the spreadsheets.