Validation intoRDF
Added validation to check if a column header is missing a value but has data present in its column; this is not allowed anymore, and an exception is thrown. Also added code to check if a cell contains an empty formula (see TestData/EmptyFormulaCells.xlsx for an example). Added a test to check that an exception is thrown if the passed sheet name is not present in the Excel