feat: support custom predicates
We would like to make some of the columns in a spreadsheet rdfs:label or
skos:definition. More generally we would like to specify a completely
custom predicate for a given column name as opposed to being limited to
a common prefix for an entire excel sheet / table.

Tried my best to be backward compatible, but will probably suggest some
renaming / refactoring in a future PR. TargetPathSegment in particular
is now a misleading name. I also find it confusing which config values
are controlling the predicate and which are controlling the object of
the transformation of a given cell into an RDF triple